+++ CHANGES.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## x.x.x (...)
+- **BREAKING** add API calls for integration with `vpn-manage-portal`, 
+  configuration file needs to be updated with new `[ApiAuthentication]` 
+  section
 - remove the `/config/` sub folder
 - remove the embedded documentation, does not belong here
 - add page to show when user is blocked